Moved the struct UuidCodec to a new file
